refactor: rename auth-fns.ts → auth-fns.server.ts
Enforces strict server-side boundary via Vite+ .server.ts convention.
Updates all import paths in login-form, signup-form, onboarding route,
and vite.config.ts server function list.
