+++ vite.config.ts
@@ -88,7 +88,7 @@ const config = defineConfig({
           "src/server/middleware/session.ts",
           "src/server/transactions.ts",
           "src/server/smart-rules.ts",
-          "src/server/auth-fns.ts",
+          "src/server/auth-fns.server.ts",
           "src/routes/api/auth/$.ts",
         ],
       },
